Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/2d.canvas.readonly-expected.txt

2d.canvas.readonly
CanvasRenderingContext2D.canvas is readonly
Actual output:

PASS CanvasRenderingContext2D.canvas is readonly
